 requirements.txt
Django==5.0
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
```

---

### **📁 Procfile** (CREATE NEW FILE in root folder)
```
web: gunicorn config.wsgi --log-file -